api/requirements.txt
black==24.8.0
requests-oauthlib==2.0.0

# For c2cwsgiutils
c2cwsgiutils[broadcast]==6.0.8
cornice==6.1.0
gunicorn==23.0.0
objgraph==3.6.1
prometheus-client==0.20.0
pyjwt==2.9.0
pyramid-mako==1.1.0
pyramid-tm==2.5
sentry-sdk==2.13.0
sqlalchemy==2.0.33
waitress==3.0.0
zope.sqlalchemy==3.1
